Split what the structure gate names
The bench separates by contract half: the vehicle adapter that settles
a law and keeps its receipts, and the campaign-facing plumbing — the
factory, the gates, the stage and its scenarios — leave the trial
physics in the parent. The scoring evaluator's objective map moves to
its own constructor, and the sealed-campaign fixture's training
prelude moves beside the fixture it opens. Code motion and extraction
only; no behavior change.
