+++ tools/flight-tune-campaign/src/bench.rs
@@ -23,17 +23,11 @@ use std::rc::Rc;
 use std::time::Duration;
 
 use flight_tune::{
-    AdapterError, ArtifactIdentity, Candidate, CandidateReceipt, CandidateTransitionReceipt,
-    CandidateTransitionRequest, Digest, EvaluatorError, GateEvaluator, GateOutcome,
-    RunExecutionContext, RunPreparationReceipt, SampleEvent, ScenarioRef, ScenarioStartReceipt,
-    SessionChallenge, SimulatorBackend, SimulatorCapability, SimulatorSessionReceipt,
-    SimulatorVehicleAdapter, SimulatorVehicleFactory, TelemetrySample, TransitionBindingReceipt,
-    VehicleBinding, VehicleBindingReceipt,
+    AdapterError, ArtifactIdentity, Candidate, Digest, RunExecutionContext, RunPreparationReceipt,
+    SampleEvent, ScenarioRef, ScenarioStartReceipt, SessionChallenge, SimulatorBackend,
+    SimulatorCapability, SimulatorSessionReceipt, TelemetrySample,
 };
-use pilotage_control_feel::{
-    AxisCurve, AxisDemandShaper, AxisDynamics, AxisResponse, FeelMode, FlightFeelProfile,
-    NeutralBand,
-};
+use pilotage_control_feel::{AxisCurve, AxisDemandShaper, AxisDynamics, AxisResponse, NeutralBand};
 
 use crate::scoring::channel;
 
@@ -167,70 +161,6 @@ fn response_from(candidate: &Candidate) -> Result<AxisResponse, AdapterError> {
             reversal_jerk: apply_jerk as f32,
         },
     })
-}
-
-/// The vehicle half of the contract: it settles a candidate's command law.
-#[derive(Debug)]
-pub struct BenchVehicleAdapter {
-    handle: BenchHandle,
-}
-
-impl BenchVehicleAdapter {
-    /// Creates one adapter over shared state.
-    #[must_use]
-    pub fn new(handle: BenchHandle) -> Self {
-        Self { handle }
-    }
-}
-
-impl SimulatorVehicleAdapter for BenchVehicleAdapter {
-    fn authorize_candidate_transition(
-        &self,
-        request: &CandidateTransitionRequest,
-    ) -> Result<CandidateTransitionReceipt, AdapterError> {
-        CandidateTransitionReceipt::authorized(request)
-            .map_err(|error| AdapterError::new(error.to_string()))
-    }
-
-    fn ensure_settled_candidate_blocking(
-        &mut self,
-        _capability: &SimulatorCapability,
-        candidate: &Candidate,
-        candidate_digest: Digest,
-    ) -> Result<CandidateReceipt, AdapterError> {
-        let mut settled = self.handle.0.borrow_mut();
-        // Repeating the request must not rewrite the law, so a restart can be
-        // reconciled without disturbing a vehicle already on the candidate.
-        if settled.digest != Some(candidate_digest) {
-            settled.response = Some(response_from(candidate)?);
-            settled.digest = Some(candidate_digest);
-        }
-        Ok(CandidateReceipt {
-            session_digest: _capability.session_digest(),
-            requested_digest: candidate_digest,
-            applied_digest: candidate_digest,
-            // The bench applies the law in process, so what it reads back is
-            // what it applied. A vehicle over a link reads its controller.
-            readback_digest: candidate_digest,
-            // Idle reconciliation settles a candidate without a run behind it.
-            run_intent_digest: None,
-        })
-    }
-
-    fn ensure_candidate_for_run_blocking(
-        &mut self,
-        capability: &SimulatorCapability,
-        context: &RunExecutionContext,
-        candidate: &Candidate,
-        candidate_digest: Digest,
-    ) -> Result<CandidateReceipt, AdapterError> {
-        let mut receipt =
-            self.ensure_settled_candidate_blocking(capability, candidate, candidate_digest)?;
-        // The receipt names the run it was settled for, so a law applied for
-        // one run cannot be read as evidence for another.
-        receipt.run_intent_digest = Some(context.digest().map_err(to_adapter)?);
-        Ok(receipt)
-    }
 }
 
 /// The simulator half: it flies the trial and reports the channels.
@@ -405,307 +335,18 @@ impl SimulatorBackend for BenchBackend {
     }
 }
 
-impl flight_tune::RunTerminalAdapter for BenchVehicleAdapter {
-    fn terminal_capabilities(&self) -> flight_tune::RunTerminalCapabilities {
-        // The bench holds no external control path, no trace collector and no
-        // supervised child: the law, the vehicle and the trace are all in this
-        // process, and a run ends when the loop stops stepping. Advertising a
-        // capability it does not have would have the engine wait for a stop
-        // acknowledgement nothing will send.
-        flight_tune::RunTerminalCapabilities::new(false, false, false)
-    }
-
-    fn bind_terminal_plan_blocking(
-        &mut self,
-        _capability: &SimulatorCapability,
-        _binding: &flight_tune::RunBindingReceipt,
-        _plan: &flight_tune::RunTerminalPlan,
-    ) -> Result<(), AdapterError> {
-        // Nothing to bind: with no external component to stop, the plan is
-        // empty and accepting it is the whole of the work. Repeating the call
-        // must be safe, and doing nothing twice is.
-        Ok(())
-    }
-
-    fn causal_evidence_digest_blocking(
-        &mut self,
-        _capability: &SimulatorCapability,
-        _binding: &flight_tune::RunBindingReceipt,
-    ) -> Result<Digest, AdapterError> {
-        // The causal evidence for an in-process run is the run itself: there
-        // is no external trace to correlate against, so the identity stated
-        // here is the one fixed value that says which trace this was.
-        Ok(Digest::from_bytes([0x0c; 32]))
-    }
-
-    fn seal_terminal_receipt_blocking(
-        &mut self,
-        _capability: &SimulatorCapability,
-        _binding: &flight_tune::RunBindingReceipt,
-        receipt: &flight_tune::RunTerminalReceipt,
-    ) -> Result<(), AdapterError> {
-        // A seal the adapter acknowledges and does not keep is a seal the
-        // engine cannot read back, and the engine is right to refuse one.
-        // Sealing the same receipt twice keeps one copy, so a repeated call
-        // after an uncertain acknowledgement is safe.
-        let mut settled = self.handle.0.borrow_mut();
-        if !settled.sealed.contains(receipt) {
-            settled.sealed.push(receipt.clone());
-        }
-        Ok(())
-    }
-
-    fn recover_terminal_receipts_blocking(
-        &mut self,
-        _capability: &SimulatorCapability,
-        _binding: &flight_tune::RunBindingReceipt,
-    ) -> Result<Vec<flight_tune::RunTerminalReceipt>, AdapterError> {
-        Ok(self.handle.0.borrow().sealed.clone())
-    }
-}
-
 /// Reads an engine error as an adapter one.
-fn to_adapter(error: flight_tune::TuneError) -> AdapterError {
+pub(super) fn to_adapter(error: flight_tune::TuneError) -> AdapterError {
     AdapterError::new(error.to_string())
 }
 
-/// Binds the bench vehicle to a validated simulator session.
-#[derive(Debug)]
-pub struct BenchVehicleFactory {
-    handle: BenchHandle,
-    identity: ArtifactIdentity,
-    transition_validator: ArtifactIdentity,
-    adjacency_policy_digest: Digest,
-}
+mod adapter;
+mod qualifying;
 
-impl BenchVehicleFactory {
-    /// Creates one factory for a named vehicle.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`AdapterError`] when an identity cannot be built.
-    pub fn new(handle: BenchHandle, vehicle_id: &str) -> Result<Self, AdapterError> {
-        let text = |value: &str| {
-            ArtifactIdentity::from_text(value, value)
-                .map_err(|error| AdapterError::new(error.to_string()))
-        };
-        Ok(Self {
-            handle,
-            identity: text(vehicle_id)?,
-            transition_validator: text("bench-transition-validator")?,
-            adjacency_policy_digest: Digest::from_bytes([8; 32]),
-        })
-    }
-}
-
-impl SimulatorVehicleFactory for BenchVehicleFactory {
-    type Adapter = BenchVehicleAdapter;
-
-    fn vehicle_identity(&self) -> &ArtifactIdentity {
-        &self.identity
-    }
-
-    fn transition_validator_identity(&self) -> &ArtifactIdentity {
-        &self.transition_validator
-    }
-
-    fn adjacency_policy_digest(&self) -> Digest {
-        self.adjacency_policy_digest
-    }
-
-    fn bind_blocking(
-        self,
-        capability: &SimulatorCapability,
-    ) -> Result<VehicleBinding<Self::Adapter>, AdapterError> {
-        let transition = TransitionBindingReceipt::new(
-            capability.session_digest(),
-            self.transition_validator,
-            self.adjacency_policy_digest,
-        )?;
-        capability.bind_vehicle_with_transition(
-            BenchVehicleAdapter::new(self.handle),
-            VehicleBindingReceipt {
-                session_digest: capability.session_digest(),
-                vehicle_digest: self.identity.digest,
-            },
-            transition,
-        )
-    }
-}
-
-/// The hard gates a bench run must not trip.
-///
-/// A gate is a refusal, not a score: a run that trips one is not a worse
-/// candidate, it is a candidate whose result means nothing. Speed and
-/// acceleration ceilings here stand for the envelope a real vehicle would be
-/// held inside.
-#[derive(Debug)]
-pub struct BenchGates {
-    identity: ArtifactIdentity,
-    maximum_speed_mps: f64,
-    maximum_acceleration_mps2: f64,
-}
-
-impl BenchGates {
-    /// Creates gates at one vehicle's ceilings.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`EvaluatorError`] when the identity cannot be built.
-    pub fn new(
-        maximum_speed_mps: f64,
-        maximum_acceleration_mps2: f64,
-    ) -> Result<Self, EvaluatorError> {
-        let identity = ArtifactIdentity::from_text("bench-envelope-gates", "bench-envelope-gates")
-            .map_err(|error| EvaluatorError::new(error.to_string()))?;
-        Ok(Self {
-            identity,
-            maximum_speed_mps,
-            maximum_acceleration_mps2,
-        })
-    }
-}
-
-impl GateEvaluator for BenchGates {
-    fn identity(&self) -> &ArtifactIdentity {
-        &self.identity
-    }
-
-    fn begin(&mut self, _scenario: &ScenarioRef) -> Result<(), EvaluatorError> {
-        Ok(())
-    }
-
-    fn evaluate(&mut self, sample: &TelemetrySample) -> Result<Vec<GateOutcome>, EvaluatorError> {
-        let read = |name: &str| sample.values.get(name).copied().unwrap_or(0.0);
-        let speed = read(channel::VELOCITY_MPS).abs();
-        let acceleration = read(channel::ACCELERATION_MPS2).abs();
-        let mut outcomes = Vec::new();
-        outcomes.push(if speed <= self.maximum_speed_mps {
-            GateOutcome::pass("envelope.speed")
-        } else {
-            GateOutcome::fail("envelope.speed", format!("{speed:.3} m/s"))
-        });
-        if !outcomes[0].passed {
-            return Ok(outcomes);
-        }
-        outcomes.push(if acceleration <= self.maximum_acceleration_mps2 {
-            GateOutcome::pass("envelope.acceleration")
-        } else {
-            GateOutcome::fail("envelope.acceleration", format!("{acceleration:.3} m/s2"))
-        });
-        Ok(outcomes)
-    }
-
-    fn finish(&mut self) -> Result<(), EvaluatorError> {
-        Ok(())
-    }
-
-    fn cancel(&mut self) -> Result<(), EvaluatorError> {
-        Ok(())
-    }
-}
-
-/// The starting command law a search departs from, as candidate parameters.
-///
-/// This is the shaped mode the vehicle would otherwise ship with, expressed in
-/// the parameters the search may move. Starting a search from a law that is
-/// already safe is what keeps the first trials from being wasted on finding
-/// one.
-#[must_use]
-pub fn warm_start_parameters(mode: FeelMode) -> BTreeMap<String, f64> {
-    let profile = FlightFeelProfile::shaped(mode);
-    let axis = profile.horizontal;
-    BTreeMap::from([
-        (
-            parameter::DEADZONE.to_owned(),
-            f64::from(axis.curve.deadzone),
-        ),
-        (
-            parameter::CENTER_EXPO.to_owned(),
-            f64::from(axis.curve.center_expo),
-        ),
-        (
-            parameter::OUTER_EXPO.to_owned(),
-            f64::from(axis.curve.outer_expo),
-        ),
-        (
-            parameter::APPLY_ACCEL.to_owned(),
-            f64::from(axis.dynamics.apply_accel),
-        ),
-        (
-            parameter::APPLY_JERK.to_owned(),
-            f64::from(axis.dynamics.apply_jerk),
-        ),
-        (
-            parameter::RELEASE_FACTOR.to_owned(),
-            f64::from(axis.dynamics.release_accel / axis.dynamics.apply_accel),
-        ),
-        (
-            parameter::NEUTRAL_ENTER.to_owned(),
-            f64::from(axis.neutral.active_enter),
-        ),
-        (
-            parameter::NEUTRAL_DWELL_MS.to_owned(),
-            f64::from(axis.neutral.dwell_ms),
-        ),
-    ])
-}
-
-/// The stage one vehicle is tuned over.
-///
-/// The allowlist is the command law's own numbers, bounded around the shaped
-/// starting point rather than around zero: a search that may propose any value
-/// spends its first trials rediscovering that a control has to be stable,
-/// which is what a warm start exists to avoid.
-#[must_use]
-pub fn bench_stage(
-    id: &str,
-    promotion: flight_tune::PromotionPolicy,
-    qualification: flight_tune::QualificationPolicy,
-) -> flight_tune::SearchStage {
-    use flight_tune::ParameterBounds;
-    let bounds = |minimum: f64, maximum: f64| ParameterBounds { minimum, maximum };
-    flight_tune::SearchStage {
-        id: id.to_owned(),
-        allowlist: BTreeMap::from([
-            (parameter::DEADZONE.to_owned(), bounds(0.02, 0.12)),
-            (parameter::CENTER_EXPO.to_owned(), bounds(0.0, 0.7)),
-            (parameter::OUTER_EXPO.to_owned(), bounds(0.0, 0.7)),
-            (parameter::APPLY_ACCEL.to_owned(), bounds(1.0, 10.0)),
-            (parameter::APPLY_JERK.to_owned(), bounds(5.0, 80.0)),
-            (parameter::RELEASE_FACTOR.to_owned(), bounds(1.0, 2.0)),
-            (parameter::NEUTRAL_ENTER.to_owned(), bounds(0.01, 0.06)),
-            (parameter::NEUTRAL_DWELL_MS.to_owned(), bounds(20.0, 200.0)),
-        ]),
-        fixed_parameters: BTreeMap::new(),
-        required_hard_gates: vec![
-            "envelope.speed".to_owned(),
-            "envelope.acceleration".to_owned(),
-        ],
-        // Training, promotion and final qualification fly separate scenarios.
-        // The bar that decides what ships is measured on runs the search never
-        // saw, so a candidate cannot be fitted to the scenario that judges it.
-        training_scenarios: vec![bench_scenario("training-step", 21)],
-        promotion_scenarios: vec![bench_scenario("promotion-step", 22)],
-        final_qualification_scenarios: vec![bench_scenario("final-step", 23)],
-        repetitions: 2,
-        promotion,
-        qualification,
-    }
-}
-
-/// One trial of the bench, as a scenario reference.
-#[must_use]
-pub fn bench_scenario(id: &str, digest_byte: u8) -> ScenarioRef {
-    ScenarioRef {
-        id: id.to_owned(),
-        digest: Digest::from_bytes([digest_byte; 32]),
-        // The trial is nine seconds at fifty hertz, and the ceiling is the
-        // sample count that covers it with room for the completion event.
-        max_samples: 700,
-        sample_timeout_ms: 200,
-    }
-}
+pub use adapter::BenchVehicleAdapter;
+pub use qualifying::{
+    BenchGates, BenchVehicleFactory, bench_scenario, bench_stage, warm_start_parameters,
+};
 
 #[cfg(test)]
 mod tests;
